ai-agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
duckdb==0.9.1
google-generativeai==0.3.0
groq==0.4.2
pandas==2.1.1
numpy==1.26.2
pyarrow==14.0.0
openpyxl==3.1.5
httpx==0.25.2
pydantic-extra-types==2.3.0
python-multipart==0.0.6
